Fix Project dependencies
Now project `$init_args` requires keys `php_version` and `wp_version` to be set.
Because `wde_replace_*` strings are replaced in `src/` but not in `vendor/`.

Keys `plugins` and `php_ext` are optional. Empty array will be used if key not exists.

### Update Info:
Add keys `php_version` and `wp_version` to `$init_args` in `functions.php` or the plugin main file.